remove old implmentation of `ObservableNode`
