+++ projects/ng-app-state/src/lib/tree-based-observable/observable-node.ts
@@ -1,19 +1,21 @@
 import { forOwn, pull } from "micro-dash";
 import { Observable, Subscriber, Subscription } from "rxjs";
+import { take } from "rxjs/operators";
 import { ObjectWith } from "s-ng-dev-utils";
 
 /** @hidden */
 export class ObservableNode extends Observable<any> {
   private value: any;
   private valueChanged = false;
-  private children: ObjectWith<ObservableNode> = {};
+  private children: ObjectWith<Set<ObservableNode>> = {};
   private subscribers: Array<Subscriber<any>> = [];
   private sourceSubscription?: Subscription;
 
-  /**
-   * @param source expected to act as if from a behavior, and only emit distinct values
-   */
-  constructor(private _source: Observable<any>, private cleanup: VoidFunction) {
+  constructor(
+    private _source: Observable<any>,
+    private key?: string,
+    private parent?: ObservableNode,
+  ) {
     super();
   }
 
@@ -32,19 +34,13 @@ export class ObservableNode extends Observable<any> {
     };
   }
 
-  ensureChild(key: string, cleanup: VoidFunction) {
-    let child = this.getChild(key);
-    if (!child) {
-      child = this.children[key] = new ObservableNode(this, cleanup);
-      if (this.value) {
-        child.value = this.value[key];
-      }
+  ensureChild(key: string) {
+    const set = this.children[key];
+    if (set && set.size) {
+      return set.values().next().value;
+    } else {
+      return new ObservableNode(this, key, this);
     }
-    return child;
-  }
-
-  removeChild(key: string) {
-    delete this.children[key];
   }
 
   updateCache(value: any) {
@@ -54,8 +50,10 @@ export class ObservableNode extends Observable<any> {
 
     this.value = value;
     this.valueChanged = true;
-    forOwn(this.children, (child, key) => {
-      child.updateCache(value ? value[key] : undefined);
+    forOwn(this.children, (children, key) => {
+      for (const child of children) {
+        child.updateCache(value ? value[key] : undefined);
+      }
     });
   }
 
@@ -63,15 +61,20 @@ export class ObservableNode extends Observable<any> {
     return this.subscribers.length === 0;
   }
 
-  getCache() {
+  getValue() {
+    if (this.subscribersAreEmpty()) {
+      if (this.parent && this.key) {
+        const parentValue: any = this.parent.getValue();
+        return parentValue ? parentValue[this.key] : undefined;
+      }
+
+      this.pipe(take(1)).subscribe();
+    }
     return this.value;
   }
 
-  private getChild(key: string) {
-    return this.children[key];
-  }
-
   private start() {
+    this.registerWithParent();
     this.sourceSubscription = this._source.subscribe(() => {
       if (this.valueChanged) {
         this.emit();
@@ -88,6 +91,35 @@ export class ObservableNode extends Observable<any> {
 
   private stop() {
     this.sourceSubscription!.unsubscribe();
-    this.cleanup();
+    if (this.parent && this.key) {
+      this.parent.unregisterChild(this.key, this);
+    }
+  }
+
+  private registerWithParent() {
+    if (this.parent && this.key) {
+      this.parent.registerChild(this.key, this);
+    }
+  }
+
+  private registerChild(key: string, child: ObservableNode) {
+    let set = this.children[key];
+    if (!set) {
+      set = this.children[key] = new Set<ObservableNode>();
+    } else if (set.has(child)) {
+      return;
+    }
+
+    set.add(child);
+    this.registerWithParent();
+    child.value = this.value ? this.value[key] : undefined;
+  }
+
+  private unregisterChild(key: string, child: ObservableNode) {
+    const set = this.children[key];
+    set.delete(child);
+    if (set.size === 0) {
+      delete this.children[key];
+    }
   }
 }
